requirements.txt
Django
gunicorn
django-heroku
djangorestframework
markdown
django-filter
